sdcsdv

+++ app/src/main/java/com/webmyne/connect/postLead/model/PostLeadInput.java
@@ -7,13 +7,13 @@ import java.util.Arrays;
  */
 public class PostLeadInput {
     public String Description, REmail, RMobile, RName;
-    public int UserID;
+    public long UserID;
     public int[] lstLeadTypeId;
 
     public PostLeadInput() {
     }
 
-    public PostLeadInput(String description, String REmail, String RMobile, String RName, int userID, int[] lstLeadTypeId) {
+    public PostLeadInput(String description, String REmail, String RMobile, String RName, long userID, int[] lstLeadTypeId) {
         this.Description = description;
         this.REmail = REmail;
         this.RMobile = RMobile;
